VS Code: Java 11 - Setup
