+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+  "configurations": [
+    {
+      "type": "java",
+      "name": "Simulation GUI",
+      "request": "launch",
+      "mainClass": "view.View",
+      "projectName": "DC2300-Coursework",
+      "vmArgs": "--module-path \"C:/Program Files/Java/javafx-sdk-11.0.2/lib\" --add-modules javafx.controls,javafx.fxml"
+    },
+    {
+      "type": "java",
+      "name": "Simulation CLI",
+      "request": "launch",
+      "mainClass": "simulation.Simulator",
+      "projectName": "DC2300-Coursework",
+      "vmArgs": "--module-path \"C:/Program Files/Java/javafx-sdk-11.0.2/lib\" --add-modules javafx.controls,javafx.fxml",
+      "args": "${input:simFile}"
+    }
+  ],
+  "inputs": [
+    {
+      "id": "simFile",
+      "description": "Location of the sim file",
+      "type": "promptString",
+      "default": "./sample-configs/twoShelves-v2.sim",
+    }
+  ]
+}
