+++ .vscode/settings.json
@@ -1,3 +1,5 @@
 {
-  "java.autobuild.enabled": false
+  "java.autobuild.enabled": true,
+  "java.dependency.packagePresentation": "hierarchical",
+  "java.home": "C:/Program Files/Java/jdk-11.0.3"
 }
