+++ src/model/Robot.java
@@ -2,6 +2,7 @@ package model;
 
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
+import utils.PathFindingStrategy;
 
 public class Robot extends Entity implements Actor {
 	private int powerUnits;
