+++ src/test/StorageShelfTest.java
@@ -1,4 +1,4 @@
-package Test;
+package test;
 
 import static org.junit.Assert.*;
 import org.junit.Test;
